+++ test/unit/fuzz-tools.test.js
@@ -406,6 +406,7 @@ import * as gellCoombsV426 from '../../lib/gell-coombs-v426.js';
 import * as vaughanWilliamsV427 from '../../lib/vaughan-williams-v427.js';
 import * as mrcPowerV428 from '../../lib/mrc-power-v428.js';
 import * as sarnatHieV429 from '../../lib/sarnat-hie-v429.js';
+import * as papileIvhV430 from '../../lib/papile-ivh-v430.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -757,6 +758,7 @@ const MODULES = {
   'vaughan-williams-v427.js': vaughanWilliamsV427,
   'mrc-power-v428.js': mrcPowerV428,
   'sarnat-hie-v429.js': sarnatHieV429,
+  'papile-ivh-v430.js': papileIvhV430,
   'tb-testing.js': tbTesting,
 };